syncc

+++ src/BaseLib/Fixed.h
@@ -7,7 +7,7 @@ typedef signed char        int8_t;
 typedef short              int16_t;
 #endif
 #ifndef int32_t
-typedef int                int32_t;
+typedef long               int32_t;
 #endif
 #ifndef int64_t
 typedef long long          int64_t;
@@ -19,7 +19,7 @@ typedef unsigned char      uint8_t;
 typedef unsigned short     uint16_t;
 #endif
 #ifndef uint32_t
-typedef unsigned int       uint32_t;
+typedef unsigned long      uint32_t;
 #endif
 #ifndef uint64_t
 typedef unsigned long long uint64_t;
